+++ app/navigators/appNavigator.js
@@ -3,7 +3,7 @@ import { createStackNavigator } from '@react-navigation/stack';
 import SplashScreen from '@scenes/SplashScreen/';
 import ExampleScreen from '@scenes/ExampleScreen';
 import { NavigationContainer } from '@react-navigation/native';
-import NavigationService from '../services/NavigationService';
+import { setTopLevelNavigator } from '@services/navigationService';
 const Stack = createStackNavigator();
 /**
  * The root screen contains the application's navigation.
@@ -12,7 +12,7 @@ const Stack = createStackNavigator();
  */
 export default function AppNavigator() {
   return (
-    <NavigationContainer ref={NavigationService.setTopLevelNavigator}>
+    <NavigationContainer ref={setTopLevelNavigator}>
       <Stack.Navigator headerMode="none" initialRouteName="SplashScreen">
         <Stack.Screen name="SplashScreen" component={SplashScreen} />
         <Stack.Screen name="MainScreen" component={ExampleScreen} />
